Extract hosted voice settings panel from settings window
